Refactor bot to improve structure and maintainability
This commit refactors the entire bot to follow a more professional and maintainable structure. The changes include:

- A new directory structure under `src/`.
- Separation of concerns into services, models, utils, and config.
- A new `Bot` class and `main.py` entry point.
- Flexible logging configuration through `config.yml`.
- Improved code quality with docstrings and type hints.
